starting to add the ai back next i will make menu then finish ai medium then I will cleaning and finally work on hard ai

+++ Node.cpp
@@ -14,7 +14,7 @@ int Node::simulate_play(const e_color& color) {
     simulation.set_color(getAction_position().first, getAction_position().second, color);
     simulation.change_color(flip_coordinates, color);
     simulation.setNumber_of_turn(simulation.getNumber_of_turn()+1);
-    return flip_coordinates.size();
+    return (int) flip_coordinates.size();
 
 }
